+++ src/common/commonTranslations.js
@@ -303,7 +303,8 @@ define('commonTranslations', [], function() {
             },
             calendar_selector: {
               'min_time': 'The collection of the previous reservation is at {{time}}',
-              'max_time': 'The delivery of the next reservation is at {{time}}'
+              'max_time': 'The delivery of the next reservation is at {{time}}',
+              'min_duration': 'min {{days}}d'
             },
         		selector: {
                 'select': 'Select',
@@ -775,7 +776,8 @@ define('commonTranslations', [], function() {
             },        
             calendar_selector: {
               'min_time': 'La devolución de la reserva anterior es a las {{time}}',
-              'max_time': 'La entrega de la siguiente reserva es a las {{time}}'
+              'max_time': 'La entrega de la siguiente reserva es a las {{time}}',
+              'min_duration': 'min {{days}}d'
             },  		
             selector: {
               'select': 'Seleccionar',              
@@ -1247,7 +1249,8 @@ define('commonTranslations', [], function() {
         },
         calendar_selector: {
           'min_time': 'La recollida de la reserva anterior és a les {{time}}',
-          'max_time': "L'entrega de la propera reserva és a les {{time}}"
+          'max_time': "L'entrega de la propera reserva és a les {{time}}",
+          'min_duration': 'min {{days}}d'
         },      
         selector: {
           'select': 'Seleccionar',          
@@ -1719,7 +1722,8 @@ define('commonTranslations', [], function() {
         },  
         calendar_selector: {
           'min_time': 'Die Abholung der vorherigen Reservierung erfolgt um {{time}} Uhr',
-          'max_time': 'Die Lieferung der nächsten Reservierung erfolgt um {{time}} Uhr'
+          'max_time': 'Die Lieferung der nächsten Reservierung erfolgt um {{time}} Uhr',
+          'min_duration': 'min {{days}}d'
         },
         selector: {
           'select': 'Selektionär',          
@@ -2192,7 +2196,8 @@ define('commonTranslations', [], function() {
         },
         calendar_selector: {
           'min_time': 'La collecte de la réservation précédente est à {{time}}',
-          'max_time': 'La livraison de la prochaine réservation est à {{time}}'
+          'max_time': 'La livraison de la prochaine réservation est à {{time}}',
+          'min_duration': 'min {{days}}d'
         },
         selector: {
           'select': 'Sélectionnez',
@@ -2664,7 +2669,8 @@ define('commonTranslations', [], function() {
         },
         calendar_selector: {
           'min_time': 'La raccolta della prenotazione precedente è alle  {{time}}',
-          'max_time': 'La consegna della prossima prenotazione è alle {{time}}'
+          'max_time': 'La consegna della prossima prenotazione è alle {{time}}',
+          'min_duration': 'min {{days}}d'
         },  
         selector: {
           'select': 'Seleziona',
@@ -2889,7 +2895,8 @@ define('commonTranslations', [], function() {
           },
           calendar_selector: {
             'min_time': 'Сбор предыдущего бронирования находится по адресу {{time}}',
-            'max_time': 'Доставка следующей брони по адресу {{time}}'
+            'max_time': 'Доставка следующей брони по адресу {{time}}',
+            'min_duration': 'мин {{days}}'
           },
             selector: {
             'select': 'селекционер',